Update WEBII-cookies.py
This is a Python3 for Web Cookie management reference sheet for CS1115. 
Note, it is not a replacement for Derek Bridge's notes.

In this example, we use a cookie to detect whether or not a user has voted before.
We assume, that the page connects to a database called 'votes' which stores two columns, band name and num_votes.
There re currently only two bands in the database. 
It would make more sense to procedurally generate the form using a loop for every row in the database, and adding the options to the select input tag, but I won't cover that yet.